backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.3.0
Werkzeug==2.0.1
requests==2.28.1
python-dotenv==0.19.2
chromadb==0.4.6
sentence-transformers==2.2.2
langchain==0.0.267
google-generativeai==0.3.0
python-docx==0.8.11
PyPDF2==3.0.1
pydantic==1.10.8
redis==4.5.1
eventlet==0.33.3
gunicorn==20.1.0
